server/requirements.txt
# django-rest-cli
Django>=3.1
djangorestframework>=3.13.1
python-decouple>=3.6
dj-database-url>=0.5.0
pytest-django>=4.5.2
dj-rest-auth
djangorestframework-simplejwt==4.7.1
django-allauth>=0.24.1
drf_spectacular
psycopg2-binary

# Populate
factory-boy==3.2.1
Faker==18.6.0
halo==0.0.31
log-symbols==0.0.14
spinners==0.0.24
termcolor==2.3.0

# Custom
python-dotenv==1.0.0
drf-standardized-errors==0.12.5
Pillow==9.5.0
django-phonenumbers==1.0.1
django-phonenumber-field==7.1.0
djangorestframework-composed-permissions==0.2.1
rest_condition==1.0.3
drf-extra-fields==3.4.1
djangorestframework-gis==1.0
django-cleanup==7.0.0
django-filter==23.2
channels==4.0.0
daphne==4.0.0